remove protocol and cleanup implementation
